Added filtering by environment to group user reports. made a test case so I could reuse test code.
