+++ src/sentry/api/endpoints/group_user_reports.py
@@ -3,10 +3,11 @@ from __future__ import absolute_import
 from sentry.api.bases.group import GroupEndpoint
 from sentry.api.serializers import serialize
 from sentry.api.paginator import DateTimePaginator
-from sentry.models import UserReport
+from sentry.models import UserReport, Environment
+from sentry.api.base import EnvironmentMixin
 
 
-class GroupUserReportsEndpoint(GroupEndpoint):
+class GroupUserReportsEndpoint(GroupEndpoint, EnvironmentMixin):
     def get(self, request, group):
         """
         List User Reports
@@ -19,7 +20,18 @@ class GroupUserReportsEndpoint(GroupEndpoint):
         :auth: required
         """
 
-        report_list = UserReport.objects.filter(group=group)
+        try:
+            environment = self._get_environment_from_request(
+                request,
+                group.organization.id,
+            )
+        except Environment.DoesNotExist:
+            report_list = Environment.objects.none()
+        else:
+            if environment is not None:
+                report_list = UserReport.objects.filter(group=group, environment=environment)
+            else:
+                report_list = UserReport.objects.filter(group=group)
 
         return self.paginate(
             request=request,
