+++ tests/sentry/api/endpoints/test_project_user_reports.py
@@ -1,10 +1,10 @@
 from __future__ import absolute_import
 
 import six
-import logging
+from exam import fixture
 
-from sentry.testutils import APITestCase
-from sentry.models import Environment, EventUser, GroupStatus, UserReport
+from sentry.testutils import APITestCase, UserReportEnvironmentTestCase
+from sentry.models import EventUser, GroupStatus, UserReport
 from sentry.event_manager import EventManager
 
 
@@ -218,66 +218,13 @@ class CreateProjectUserReportTest(APITestCase):
         assert euser.name == 'Foo Bar'
 
 
-class ProjectUserReportByEnvironmentsTest(APITestCase):
-    def setUp(self):
-
-        self.project = self.create_project()
-        self.env1 = self.create_environment(self.project, 'production')
-        self.env2 = self.create_environment(self.project, 'staging')
-
-        self.group = self.create_group(project=self.project)
-
-        self.env1_events = self.create_events_for_environment(self.group, self.env1, 5)
-        self.env2_events = self.create_events_for_environment(self.group, self.env2, 5)
-
-        self.env1_userreports = self.create_user_report_for_events(
-            self.project, self.group, self.env1_events, self.env1)
-        self.env2_userreports = self.create_user_report_for_events(
-            self.project, self.group, self.env2_events, self.env2)
-
-        self.path = '/api/0/projects/{}/{}/user-feedback/'.format(
+class ProjectUserReportByEnvironmentsTest(UserReportEnvironmentTestCase):
+    @fixture
+    def path(self):
+        return '/api/0/projects/{}/{}/user-feedback/'.format(
             self.project.organization.slug,
             self.project.slug,
         )
-
-    def make_event(self, **kwargs):
-        result = {
-            'event_id': 'a' * 32,
-            'message': 'foo',
-            'timestamp': 1403007314.570599,
-            'level': logging.ERROR,
-            'logger': 'default',
-            'tags': [],
-        }
-        result.update(kwargs)
-        return result
-
-    def create_environment(self, project, name):
-        env = Environment.objects.create(
-            project_id=project.id,
-            organization_id=project.organization_id,
-            name=name,
-        )
-        env.add_project(project)
-        return env
-
-    def create_events_for_environment(self, group, environment, num_events):
-        return [self.create_event(group=group, tags={
-            'environment': environment.name}) for __i in range(num_events)]
-
-    def create_user_report_for_events(self, project, group, events, environment):
-        reports = []
-        for i, event in enumerate(events):
-            reports.append(UserReport.objects.create(
-                group=group,
-                project=project,
-                event_id=event.event_id,
-                name='foo%d' % i,
-                email='bar%d@example.com' % i,
-                comments='It Broke!!!',
-                environment=environment,
-            ))
-        return reports
 
     def test_environment_gets_user_report(self):
         event_id = 'a' * 32
@@ -326,12 +273,6 @@ class ProjectUserReportByEnvironmentsTest(APITestCase):
         assert response.status_code == 200, response.content
         assert UserReport.objects.get(event_id=event_id).environment == self.env1
 
-    def assert_same_userreports(self, response_data, userreports):
-        assert sorted(int(r.get('id')) for r in response_data) == sorted(
-            r.id for r in userreports)
-        assert sorted(r.get('eventID') for r in response_data) == sorted(
-            r.event_id for r in userreports)
-
     def test_specified_enviroment(self):
         self.login_as(user=self.user)
 
@@ -348,8 +289,8 @@ class ProjectUserReportByEnvironmentsTest(APITestCase):
     def test_no_environment_does_not_exists(self):
         self.login_as(user=self.user)
         response = self.client.get(self.path + '?environment=')
-        assert response.status_code == 400
-        assert response.data == {'environment': 'Invalid environment'}
+        assert response.status_code == 200
+        assert response.data == []
 
     def test_no_environment(self):
         self.login_as(user=self.user)
@@ -376,5 +317,5 @@ class ProjectUserReportByEnvironmentsTest(APITestCase):
     def test_invalid_environment(self):
         self.login_as(user=self.user)
         response = self.client.get(self.path + '?environment=invalid_env')
-        assert response.status_code == 400
-        assert response.data == {'environment': 'Invalid environment'}
+        assert response.status_code == 200
+        assert response.data == []
